Move non-public env variable to `build-env.ts` so it doesn't pollute and throw error when other imports are used
